refactor(FormattedDuration): Make default export, rename file.
