+++ lib/components/narrative/default/default-itinerary.js
@@ -8,7 +8,7 @@ import FieldTripGroupSize from '../../admin/field-trip-itinerary-group-size'
 import NarrativeItinerary from '../narrative-itinerary'
 import ItineraryBody from '../line-itin/connected-itinerary-body'
 import SimpleRealtimeAnnotation from '../simple-realtime-annotation'
-import { FormattedDuration } from '../../util/format-duration'
+import FormattedDuration from '../../util/formatted-duration'
 import FormattedTimeRange from '../../util/formatted-time-range'
 
 import ItinerarySummary from './itinerary-summary'
