+++ lib/components/narrative/line-itin/itin-summary.js
@@ -6,7 +6,7 @@ import { connect } from 'react-redux'
 import { FormattedNumber, FormattedMessage } from 'react-intl'
 
 import { ComponentContext } from '../../../util/contexts'
-import { FormattedDuration } from '../../util/format-duration'
+import FormattedDuration from '../../util/formatted-duration'
 import FormattedTimeRange from '../../util/formatted-time-range'
 
 // TODO: make this a prop
