+++ lib/components/narrative/realtime-annotation.js
@@ -3,7 +3,7 @@ import React, { Component } from 'react'
 import { Button, OverlayTrigger, Popover } from 'react-bootstrap'
 import { FormattedList, FormattedMessage } from 'react-intl'
 
-import { FormattedDuration } from '../util/format-duration'
+import FormattedDuration from '../util/formatted-duration'
 import Icon from '../util/icon'
 
 export default class RealtimeAnnotation extends Component {
