+++ lib/components/narrative/tabbed-itineraries.js
@@ -10,7 +10,7 @@ import * as narrativeActions from '../../actions/narrative'
 import { ComponentContext } from '../../util/contexts'
 import { getTimeFormat } from '../../util/i18n'
 import { getActiveSearch, getRealtimeEffects } from '../../util/state'
-import { FormattedDuration } from '../util/format-duration'
+import FormattedDuration from '../util/formatted-duration'
 import FormattedTimeRange from '../util/formatted-time-range'
 
 const { calculateFares, calculatePhysicalActivity } = coreUtils.itinerary
