+++ lib/components/util/formatted-duration.js
@@ -4,7 +4,7 @@ import { FormattedMessage } from 'react-intl'
 /**
  * Formats the given duration according to the selected locale.
  */
-export function FormattedDuration ({duration}) {
+export default function FormattedDuration ({duration}) {
   const dur = moment.duration(duration, 'seconds')
   const hours = dur.hours()
   const minutes = dur.minutes()
